requirements.txt
asgiref==3.6.0
boto3==1.26.113
botocore==1.29.113
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-filter==23.1
django-js-asset==2.0.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.12
hvac==1.1.0
idna==3.4
importlib-metadata==6.3.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.2
Pillow==9.5.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
win32-setctime==1.1.0
zipp==3.15.0
